requirements.txt
# Core gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Configuration
pyyaml==6.0.1

# Terminal colors and UI
colorama==0.4.6

# Process management (for stop_cluster.py)
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: for better logging
python-json-logger==2.0.7